BIGDECIMAL 처리 테스트 파일 ADD

+++ src/main/java/com/pattern/strategypattern/DiscountStrategy.java
@@ -1,9 +1,11 @@
 package com.pattern.strategypattern;
 
+import java.math.BigDecimal;
+
 public interface DiscountStrategy {
     /**
      * @param originPrice 원가(정수, 예: 10000)
      * @return 할인 금액(정수)
      */
-    int discount(int originPrice);
+    BigDecimal discount(BigDecimal originPrice);
 }
